Validation callback does not run on an empty field
I have had to make this tweak in this class on a few sites now where we have written validation rules which check that at least X number of fields have been filled in but some can be left blank. The original code does not run because in the validation class in the piklist core it checks that the field has a value first. I don't think this is needed. 

Please see our earlier conversation on this some time ago..

https://piklist.com/support/topic/validate-empty-field/

Great to see this now on Github so we can put forward propose tweaks!